+++ CHANGELOG.md
@@ -64,6 +64,7 @@
 * Python: Added RANDOMKEY command ([#1701](https://github.com/aws/glide-for-redis/pull/1701))
 * Python: Added FUNCTION FLUSH command ([#1700](https://github.com/aws/glide-for-redis/pull/1700))
 * Python: Added FUNCTION DELETE command ([#1714](https://github.com/aws/glide-for-redis/pull/1714))
+* Python: Added FUNCTION LIST command ([#1738](https://github.com/aws/glide-for-redis/pull/1738))
 * Python: Added SSCAN command ([#1709](https://github.com/aws/glide-for-redis/pull/1709))
 * Python: Added LCS command ([#1716](https://github.com/aws/glide-for-redis/pull/1716))
 * Python: Added WAIT command ([#1710](https://github.com/aws/glide-for-redis/pull/1710))